+++ src/vs/workbench/parts/extensions/common/vsoGalleryService.ts
